Reworked social buttons to handle mobile breakpoint sizing better
Fixed issue with burger menu disappearing at dark theme
