+++ src/components/common/Socials/social.json
@@ -16,5 +16,11 @@
     "name": "Twitter",
     "link": "https://twitter.com/keithrpotempa",
     "icon": "twitter"
+  },
+  {
+    "id": 4,
+    "name": "Resume",
+    "link": "https://res.cloudinary.com/krp/image/upload/v1614724673/keithpotempa.com/Potempa_Dev_Resume_v9.pdf",
+    "icon": ""
   }
 ]
