+++ src/components/theme/Header/Hamburger/styles.js
@@ -34,7 +34,6 @@ export const Bar = styled.div`
 	width: 1.6rem;
 	height: .15rem;
 	margin-bottom: .3rem;
-	background-color: #212121;
 	transition: transform 500ms cubic-bezier(0.6, 0.05, 0.28, 0.91),
 	opacity 500ms,
 	box-shadow 250ms,
@@ -43,6 +42,12 @@ export const Bar = styled.div`
 	@media (max-width: 600px){
 		width: 1.6rem;
 	}
+
+	${({ theme }) => 
+		`
+		background-color: ${(theme === 'light' ? '#212121' : '#fff')};
+		
+	`}
 
 	${({ top, sidebar, theme }) =>
     top &&
